requirements.txt
streamlit==1.26.0
pandas==2.0.3
Pillow==10.0.0 